k8_operators/py_flex_gbs_bt/requirements.txt
apache-beam[gcp]==2.22.0
google-cloud-bigquery-storage==1.0.0